move inlist rule to expr_simplifier
